% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fitmod}
\alias{fitmod}
\title{Fit Bayesian Dose Response Model Using rJAGS}
\usage{
fitmod(dat, modtext, modvars = c("prob"), burn, mcmcSamps = 1000,
  chains = 1, type = "bin", trace_plots = FALSE)
}
\arguments{
\item{dat}{trial data_frame}

\item{modtext}{BUGS formatted text of model compatibale with analysis data}

\item{modvars}{vector of model variables to sample from model}

\item{burn}{number of burn in samples}

\item{mcmcSamps}{number of MCMC samples for inference}

\item{chains}{number of sample chains}

\item{type}{response data type c("bin","cont")}

\item{trace_plots}{indicator if traceplots should be output}
}
\value{
MCMC samples (as MCMC list) from modtext on modvars
}
\description{
Fit Bayesian Dose Response Model Using rJAGS
}
\examples{
Fit bayesian sigmoid EMAX model:

fitmod(sample_data,modtext="sigEmax",modvars="dr_prob",burn=10000,mcmcSamps=10000,chains=3)

}
