% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{gendr_emax}
\alias{gendr_emax}
\title{Generate random dose-response data from sigmoid EMAX model, monotherapy (add combination next)}
\usage{
gendr_emax(type = "binomial", dose = c(0, 2, 5, 10, 15, 30),
  n = rep(50, 6), e0 = -2, emax = 2, ed50 = 7.5, hill = 2)
}
\arguments{
\item{type}{data type, only supports "binomial" currently}

\item{dose}{vector of monotherapy investigations doses}

\item{n}{vector of sample sizes by dose}

\item{e0}{placebo effect in EMAX model, logit scale for binomial data}

\item{emax}{emax effect in EMAX model, logit scale for binomial data}

\item{ed50}{ed50 parameter in EMAX model}

\item{hill}{slope parameter in EMAX model}
}
\value{
tibble of dose response data
}
\description{
Generate random dose-response data from sigmoid EMAX model, monotherapy (add combination next)
}
