% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{gendr_linearcomb}
\alias{gendr_linearcomb}
\title{Generate random combination dose response data using linear models}
\usage{
gendr_linearcomb(type = "binomial", dose_1 = c(0, 2, 5, 10, 15, 30),
  dose_2 = c(0, 1, 2, 4), n_total = 240, e0 = -2.5, theta1 = 0.25,
  theta2 = 0.5, theta12 = 0)
}
\arguments{
\item{type}{data type, only supports "binomial" currently}

\item{dose_1}{vector of monotherapy investigational doses for drug 1}

\item{dose_2}{vector of monotherapy investigational doses for drug 2}

\item{n_total}{total number of subjects, (currently) assumes equivalent number of subjects per 
dose combination}

\item{e0}{placebo effect, logit scale for binomial data}

\item{theta1}{monotherapy slope for dose1, logit scale for binomial data}

\item{theta2}{monotherapy slope for dose 2, logit scale for binomial data}

\item{theta12}{interaction term, < 0 for antagonism, 0 for additivity, >0 for synergism}
}
\value{
tibble of dose response data
}
\description{
Generate random combination dose response data using linear models
}
