% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{linear}
\alias{linear}
\title{Creates BUGS model for linear dose combination model}
\usage{
linear(datatype, prior = list(theta0 = list(pmean, psd), theta1 =
  list(pmean, psd), theta2 = list(pmean, psd), theta12 = list(pmean, psd)))
}
\arguments{
\item{datatype}{character variable for response data type, args = c(\strong{"binomial"})}

\item{prior}{list of parameters in model.  For this model, must contain lists: \strong{theta0}, 
\strong{theta1}, \strong{theta2}, \strong{theta12}.  Each list has values \emph{pmean} 
and \emph{psd} for prior means and standard deviations respectively.}
}
\description{
Creates BUGS model for linear dose combination model
}
\examples{
Dichotomous linear model:
linear("binomial", prior = list(theta0 = list(0,10), theta1 = list(0,10), 
       theta2 = list(0,10), theta12 = list(0,10)))
       
}
