% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{linearcomb}
\alias{linearcomb}
\title{Linear combination function}
\usage{
linearcomb(type, dose_1, dose_2, e0, theta1, theta2, theta12)
}
\arguments{
\item{type}{data type, "binomial" or "continuous"}

\item{dose_1}{combination dose 1}

\item{dose_2}{combination dose 2}

\item{e0}{placebo effect, logit scale for binomial data}

\item{theta1}{monotherapy slope for dose1, logit scale for binomial data}

\item{theta2}{monotherapy slope for dose 2, logit scale for binomial data}

\item{theta12}{interaction term, < 0 for antagonism, 0 for additivity, >0 for synergism}
}
\value{
single combination probability
}
\description{
Linear combination function
}
